Fix bugs
Fixed a bunch of small bugs that made it so twig could not output properly. This iteration still outputs directly on the HTML page, since I could not find a way to get the output string from the render function.
